Add crossOrigin=anonymous to video element
Adds crossOrigin=anonymous to video element in service worker sample.
crossOrigin provides CORS support and defines how the element handlers
cross-origin requests.

Setting the value to anonymous allows the Service Worker to cache the
response. Without this the Service Worker will not be able to intercept
responses as to cache them.