Gmicros/fix basic custom sample (#5)
* Add crossOrigin=anonymous to video element

Adds crossOrigin=anonymous to video element in service worker sample.
crossOrigin provides CORS support and defines how the element handlers
cross-origin requests.

Setting the value to anonymous allows the Service Worker to cache the
response. Without this the Service Worker will not be able to intercept
responses as to cache them.

* Update README for basic-custom-stream

Updates the README with instructions on how to run the
basic-custom-stream sample.

The details contain how to generate a certificate and run the server
over https.